@@ -1,17 +1,54 @@
 source 'http://rubygems.org'
 
-gem 'rails', '~> 3.2.12'
-gem 'rake', '~> 10.0.3'
-gem 'rdoc', '~> 3.12.2'
-gem 'haml', '~> 3.1.7'
 
-# Bundle edge Rails instead:
-# gem 'rails',     :git => 'git://github.com/rails/rails.git'
+# Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
+gem 'rails', '4.1.4'
 
-gem 'mysql2'
+# Use sqlite3 as the database for Active Record
+gem 'sqlite3'
 
-gem 'json'
+# Use MySQL as the database for Active Record
+#gem 'mysql2'
+
+# Use SCSS for stylesheets
+gem 'sass-rails', '~> 4.0.3'
+# Use Uglifier as compressor for JavaScript assets
+gem 'uglifier', '>= 1.3.0'
+# Use CoffeeScript for .js.coffee assets and views
+gem 'coffee-rails', '~> 4.0.0'
+# See https://github.com/sstephenson/execjs#readme for more supported runtimes
+# gem 'therubyracer',  platforms: :ruby
+
+# Use jquery as the JavaScript library
+gem 'jquery-rails'
+gem 'jquery-ui-rails'
+
+# Turbolinks makes following links in your web application faster. Read more: https://github.com/rails/turbolinks
+gem 'turbolinks'
+# Build JSON APIs with ease. Read more: https://github.com/rails/jbuilder
+gem 'jbuilder', '~> 2.0'
+# bundle exec rake doc:rails generates the API under doc/api.
+gem 'sdoc', '~> 0.4.0',          group: :doc
+
+# Spring speeds up development by keeping your application running in the background. Read more: https://github.com/rails/spring
+gem 'spring',        group: :development
+
+# Use ActiveModel has_secure_password
+# gem 'bcrypt', '~> 3.1.7'
+
+# Use unicorn as the app server
+# gem 'unicorn'
+
+# Use Capistrano for deployment
+# gem 'capistrano-rails', group: :development
+
+# Use debugger
+# gem 'debugger', group: [:development, :test]
+
+#gem 'json'
 #gem 'system_timer'
+
+gem 'haml', '4.0.5'
 
 # Gems used for mocking and stubbing while running tests
 group :test do
@@ -24,34 +61,11 @@ group :test do
 	gem 'ruby-prof'
 end
 
-# Gems used only for assets and not required
-# in production environments by default.
-group :assets do
-  gem 'sass-rails'
-  gem 'coffee-rails', '~> 3.2.1'
-  gem 'uglifier',     '>= 1.0.3'
-end
-
-gem 'jquery-rails'
-gem 'jquery-ui-rails'
-
-# To use ActiveModel has_secure_password
-# gem 'bcrypt-ruby', '~> 3.0.0'
-
-# Use unicorn as the web server
-# gem 'unicorn'
-
-# Deploy with Capistrano
-# gem 'capistrano'
-
-# To use debugger
-# gem 'ruby-debug'
-
 # Use Devise for user authentication
-gem 'devise', :git => "https://github.com/plataformatec/devise.git"
+gem 'devise'
 
 # Use Hanna as documentation template
-gem 'hanna-nouveau'
+gem 'hanna-nouveau'#, '0.3.0'
 
 # Use OmniAuth for authentication with other services
 # (we don't use Devise's built-in since it didn't work when we tried)
@@ -70,9 +84,7 @@ gem "omniauth-yahoo"
 gem "omniauth-yandex"
 gem "omniauth-youtube"
 
-gem "omniauth", '~> 1.1.4'
-#gem "oa-oauth"
-#gem "rest-client"
+gem "omniauth"
 
 # We use GeoIP so we can adjust locale depending on origin
 gem 'geoip'
@@ -81,13 +93,13 @@ gem 'geoip'
 gem 'recaptcha', :require => 'recaptcha/rails'
 
 # Let third parties access our services via OAuth
-gem "oauth-plugin", ">= 0.4.0.pre1"
+gem "oauth-plugin"
 
 # Push changes to connected clients
 gem "juggernaut"
 
 # Translate the URLs
-gem "i18n_routing"
+#gem "i18n_routing"
 
 # Paginate
 gem "will_paginate"
