@@ -9,18 +9,18 @@
 # from scratch. The latter is a flawed and unsustainable approach (the more migrations
 # you'll amass, the slower it'll run and the greater likelihood for issues).
 #
-# It's strongly recommended to check this file into your version control system.
+# It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(:version => 20140419104846) do
+ActiveRecord::Schema.define(version: 20140722151900) do
 
-  create_table "admin_configs", :force => true do |t|
+  create_table "admin_configs", force: true do |t|
     t.string   "name"
-    t.integer  "pending_donations_limit"
+    t.integer  "pending_donations_limit", default: 10
     t.datetime "created_at"
     t.datetime "updated_at"
   end
 
-  create_table "admin_translatees", :force => true do |t|
+  create_table "admin_translatees", force: true do |t|
     t.string   "name"
     t.text     "description"
     t.datetime "created_at"
@@ -28,12 +28,12 @@ ActiveRecord::Schema.define(:version => 20140419104846) do
     t.string   "size"
   end
 
-  create_table "admin_translatees_admin_translatee_params", :id => false, :force => true do |t|
+  create_table "admin_translatees_admin_translatee_params", id: false, force: true do |t|
     t.integer "admin_translatee_id"
     t.integer "admin_translatee_param_id"
   end
 
-  create_table "admin_translation_parameters", :force => true do |t|
+  create_table "admin_translation_parameters", force: true do |t|
     t.string   "name"
     t.string   "description"
     t.string   "example"
@@ -41,7 +41,7 @@ ActiveRecord::Schema.define(:version => 20140419104846) do
     t.datetime "updated_at"
   end
 
-  create_table "albums", :force => true do |t|
+  create_table "albums", force: true do |t|
     t.string   "title"
     t.integer  "year"
     t.string   "description"
@@ -50,27 +50,27 @@ ActiveRecord::Schema.define(:version => 20140419104846) do
     t.string   "art_url"
   end
 
-  create_table "albums_artists", :id => false, :force => true do |t|
-    t.integer "artist_id", :null => false
-    t.integer "album_id",  :null => false
-  end
-
-  add_index "albums_artists", ["artist_id", "album_id"], :name => "by_album_and_artist", :unique => true
-
-  create_table "albums_songs", :id => false, :force => true do |t|
+  create_table "albums_artists", id: false, force: true do |t|
     t.integer "album_id"
-    t.integer "song_id"
+    t.integer "artist_id"
   end
 
-  add_index "albums_songs", ["album_id", "song_id"], :name => "by_album_and_song", :unique => true
+  add_index "albums_artists", ["artist_id", "album_id"], name: "by_album_and_artist", unique: true
 
-  create_table "artists", :force => true do |t|
+  create_table "albums_songs", id: false, force: true do |t|
+    t.integer "song_id"
+    t.integer "album_id"
+  end
+
+  add_index "albums_songs", ["album_id", "song_id"], name: "by_album_and_song", unique: true
+
+  create_table "artists", force: true do |t|
     t.string   "name"
+    t.string   "description"
     t.datetime "created_at"
     t.datetime "updated_at"
     t.string   "picture"
-    t.text     "description"
-    t.string   "donatable_status", :default => "ok"
+    t.string   "donatable_status", default: "ok"
     t.string   "facebook"
     t.string   "twitter"
     t.string   "googleplus"
@@ -82,22 +82,22 @@ ActiveRecord::Schema.define(:version => 20140419104846) do
     t.string   "lastfm"
   end
 
-  add_index "artists", ["name"], :name => "by_name", :unique => true
+  add_index "artists", ["name"], name: "by_name", unique: true
 
-  create_table "artists_songs", :id => false, :force => true do |t|
+  create_table "artists_songs", id: false, force: true do |t|
     t.integer "artist_id"
     t.integer "song_id"
   end
 
-  add_index "artists_songs", ["artist_id", "song_id"], :name => "by_artist_and_song", :unique => true
+  add_index "artists_songs", ["artist_id", "song_id"], name: "by_artist_and_song", unique: true
 
-  create_table "client_applications", :force => true do |t|
+  create_table "client_applications", force: true do |t|
     t.string   "name"
-    t.string   "website"
+    t.string   "url"
     t.string   "support_url"
     t.string   "callback_url"
-    t.string   "key",          :limit => 40
-    t.string   "secret",       :limit => 40
+    t.string   "key",          limit: 40
+    t.string   "secret",       limit: 40
     t.integer  "user_id"
     t.datetime "created_at"
     t.datetime "updated_at"
@@ -108,9 +108,9 @@ ActiveRecord::Schema.define(:version => 20140419104846) do
     t.string   "author_url"
   end
 
-  add_index "client_applications", ["key"], :name => "index_client_applications_on_key", :unique => true
+  add_index "client_applications", ["key"], name: "index_client_applications_on_key", unique: true
 
-  create_table "column_sorts", :force => true do |t|
+  create_table "column_sorts", force: true do |t|
     t.integer  "user_id"
     t.integer  "column_id"
     t.string   "field"
@@ -119,7 +119,7 @@ ActiveRecord::Schema.define(:version => 20140419104846) do
     t.datetime "updated_at"
   end
 
-  create_table "columns", :force => true do |t|
+  create_table "columns", force: true do |t|
     t.integer  "user_id"
     t.integer  "list_config_id"
     t.string   "name"
@@ -135,7 +135,7 @@ ActiveRecord::Schema.define(:version => 20140419104846) do
     t.datetime "updated_at"
   end
 
-  create_table "configurations", :force => true do |t|
+  create_table "configurations", force: true do |t|
     t.integer  "user_id"
     t.string   "name"
     t.string   "media_state"
@@ -161,35 +161,35 @@ ActiveRecord::Schema.define(:version => 20140419104846) do
     t.datetime "updated_at"
   end
 
-  create_table "devices", :force => true do |t|
+  create_table "devices", force: true do |t|
     t.string   "name"
     t.integer  "user_id"
     t.string   "last_ip"
     t.string   "version"
     t.datetime "created_at"
     t.datetime "updated_at"
-    t.integer  "app_id"
+    t.integer  "client_application_id"
     t.integer  "configuration_id"
-    t.string   "status",           :default => "offline"
-    t.string   "channels",         :default => ""
+    t.string   "status",                default: "offline"
+    t.string   "channels",              default: ""
   end
 
-  create_table "donations", :force => true do |t|
+  create_table "donations", force: true do |t|
     t.integer  "artist_id"
-    t.decimal  "artist_percentage",  :precision => 10, :scale => 0
-    t.decimal  "stoffi_percentage",  :precision => 10, :scale => 0
-    t.decimal  "charity_percentage", :precision => 10, :scale => 0
-    t.decimal  "amount",             :precision => 10, :scale => 0
+    t.decimal  "artist_percentage"
+    t.decimal  "stoffi_percentage"
+    t.decimal  "charity_percentage"
+    t.decimal  "amount"
     t.integer  "user_id"
     t.datetime "created_at"
     t.datetime "updated_at"
     t.string   "email"
-    t.integer  "return_policy",                                     :default => 0
+    t.integer  "return_policy",      default: 0
     t.string   "message"
-    t.string   "status",                                            :default => "pending"
+    t.string   "status",             default: "pending"
   end
 
-  create_table "downloads", :force => true do |t|
+  create_table "downloads", force: true do |t|
     t.string   "ip"
     t.string   "channel"
     t.string   "arch"
@@ -198,7 +198,7 @@ ActiveRecord::Schema.define(:version => 20140419104846) do
     t.datetime "updated_at"
   end
 
-  create_table "equalizer_profiles", :force => true do |t|
+  create_table "equalizer_profiles", force: true do |t|
     t.string   "name"
     t.boolean  "is_protected"
     t.string   "levels"
@@ -208,18 +208,18 @@ ActiveRecord::Schema.define(:version => 20140419104846) do
     t.datetime "updated_at"
   end
 
-  create_table "histories", :force => true do |t|
+  create_table "histories", force: true do |t|
     t.integer  "user_id"
     t.datetime "created_at"
     t.datetime "updated_at"
   end
 
-  create_table "histories_songs", :id => false, :force => true do |t|
+  create_table "histories_songs", id: false, force: true do |t|
     t.integer "history_id"
     t.integer "song_id"
   end
 
-  create_table "keyboard_shortcut_profiles", :force => true do |t|
+  create_table "keyboard_shortcut_profiles", force: true do |t|
     t.string   "name"
     t.boolean  "is_protected"
     t.integer  "configuration_id"
@@ -227,7 +227,7 @@ ActiveRecord::Schema.define(:version => 20140419104846) do
     t.datetime "updated_at"
   end
 
-  create_table "keyboard_shortcuts", :force => true do |t|
+  create_table "keyboard_shortcuts", force: true do |t|
     t.integer  "user_id"
     t.string   "name"
     t.string   "category"
@@ -238,7 +238,7 @@ ActiveRecord::Schema.define(:version => 20140419104846) do
     t.datetime "updated_at"
   end
 
-  create_table "languages", :force => true do |t|
+  create_table "languages", force: true do |t|
     t.string   "native_name"
     t.string   "english_name"
     t.string   "iso_tag"
@@ -247,34 +247,32 @@ ActiveRecord::Schema.define(:version => 20140419104846) do
     t.datetime "updated_at"
   end
 
-  create_table "link_backlogs", :force => true do |t|
+  create_table "link_backlogs", force: true do |t|
     t.integer  "link_id"
     t.integer  "resource_id"
     t.string   "resource_type"
     t.string   "error"
-    t.datetime "created_at",    :null => false
-    t.datetime "updated_at",    :null => false
+    t.datetime "created_at"
+    t.datetime "updated_at"
   end
 
-  create_table "links", :force => true do |t|
+  create_table "links", force: true do |t|
     t.integer  "user_id"
     t.string   "provider"
     t.string   "uid"
-    t.boolean  "do_share"
-    t.boolean  "do_listen"
+    t.boolean  "do_share",           default: true
+    t.boolean  "do_listen",          default: true
     t.datetime "created_at"
     t.datetime "updated_at"
-    t.boolean  "do_donate",           :default => true
-    t.string   "access_token_secret"
-    t.string   "access_token"
-    t.boolean  "do_create_playlist",  :default => true
-    t.boolean  "show_button",         :default => true
+    t.boolean  "do_donate",          default: true
+    t.boolean  "do_create_playlist", default: true
+    t.boolean  "show_button",        default: true
     t.string   "refresh_token"
     t.datetime "token_expires_at"
     t.string   "encrypted_uid"
   end
 
-  create_table "list_configs", :force => true do |t|
+  create_table "list_configs", force: true do |t|
     t.integer  "user_id"
     t.string   "selected_indices"
     t.string   "filter"
@@ -288,7 +286,7 @@ ActiveRecord::Schema.define(:version => 20140419104846) do
     t.datetime "updated_at"
   end
 
-  create_table "listens", :force => true do |t|
+  create_table "listens", force: true do |t|
     t.integer  "user_id"
     t.integer  "song_id"
     t.integer  "playlist_id"
@@ -301,23 +299,23 @@ ActiveRecord::Schema.define(:version => 20140419104846) do
     t.datetime "started_at"
   end
 
-  create_table "oauth_nonces", :force => true do |t|
+  create_table "oauth_nonces", force: true do |t|
     t.string   "nonce"
     t.integer  "timestamp"
     t.datetime "created_at"
     t.datetime "updated_at"
   end
 
-  add_index "oauth_nonces", ["nonce", "timestamp"], :name => "index_oauth_nonces_on_nonce_and_timestamp", :unique => true
+  add_index "oauth_nonces", ["nonce", "timestamp"], name: "index_oauth_nonces_on_nonce_and_timestamp", unique: true
 
-  create_table "oauth_tokens", :force => true do |t|
+  create_table "oauth_tokens", force: true do |t|
     t.integer  "user_id"
-    t.string   "type",                  :limit => 20
+    t.string   "type",                  limit: 20
     t.integer  "client_application_id"
-    t.string   "token",                 :limit => 40
-    t.string   "secret",                :limit => 40
+    t.string   "token",                 limit: 40
+    t.string   "secret",                limit: 40
     t.string   "callback_url"
-    t.string   "verifier",              :limit => 20
+    t.string   "verifier",              limit: 20
     t.string   "scope"
     t.datetime "authorized_at"
     t.datetime "invalidated_at"
@@ -326,44 +324,44 @@ ActiveRecord::Schema.define(:version => 20140419104846) do
     t.datetime "updated_at"
   end
 
-  add_index "oauth_tokens", ["token"], :name => "index_oauth_tokens_on_token", :unique => true
+  add_index "oauth_tokens", ["token"], name: "index_oauth_tokens_on_token", unique: true
 
-  create_table "playlist_subscribers", :id => false, :force => true do |t|
+  create_table "playlist_subscribers", id: false, force: true do |t|
     t.integer "playlist_id"
     t.integer "user_id"
   end
 
-  add_index "playlist_subscribers", ["user_id", "playlist_id"], :name => "by_user_and_playlist", :unique => true
+  add_index "playlist_subscribers", ["user_id", "playlist_id"], name: "by_user_and_playlist", unique: true
 
-  create_table "playlists", :force => true do |t|
+  create_table "playlists", force: true do |t|
     t.string   "name"
     t.integer  "user_id"
-    t.boolean  "is_public",  :default => true
+    t.boolean  "is_public",  default: true
     t.datetime "created_at"
     t.datetime "updated_at"
   end
 
-  add_index "playlists", ["user_id", "name"], :name => "by_user_and_name", :unique => true
+  add_index "playlists", ["user_id", "name"], name: "by_user_and_name", unique: true
 
-  create_table "playlists_songs", :id => false, :force => true do |t|
+  create_table "playlists_songs", id: false, force: true do |t|
     t.integer "playlist_id"
     t.integer "song_id"
   end
 
-  add_index "playlists_songs", ["playlist_id", "song_id"], :name => "by_playlist_and_song", :unique => true
+  add_index "playlists_songs", ["playlist_id", "song_id"], name: "by_playlist_and_song", unique: true
 
-  create_table "queues", :force => true do |t|
+  create_table "queues", force: true do |t|
     t.integer  "user_id"
     t.datetime "created_at"
     t.datetime "updated_at"
   end
 
-  create_table "queues_songs", :id => false, :force => true do |t|
+  create_table "queues_songs", id: false, force: true do |t|
     t.integer "queue_id"
     t.integer "song_id"
   end
 
-  create_table "shares", :force => true do |t|
+  create_table "shares", force: true do |t|
     t.integer  "user_id"
     t.integer  "playlist_id"
     t.string   "message"
@@ -374,18 +372,18 @@ ActiveRecord::Schema.define(:version => 20140419104846) do
     t.string   "resource_type"
   end
 
-  create_table "song_relations", :force => true do |t|
+  create_table "song_relations", force: true do |t|
     t.integer "song1_id"
     t.integer "song2_id"
     t.integer "user_id"
     t.integer "weight"
   end
 
-  add_index "song_relations", ["song1_id"], :name => "index_song_relations_on_song1_id"
-  add_index "song_relations", ["song2_id"], :name => "index_song_relations_on_song2_id"
-  add_index "song_relations", ["user_id"], :name => "index_song_relations_on_user_id"
+  add_index "song_relations", ["song1_id"], name: "index_song_relations_on_song1_id"
+  add_index "song_relations", ["song2_id"], name: "index_song_relations_on_song2_id"
+  add_index "song_relations", ["user_id"], name: "index_song_relations_on_user_id"
 
-  create_table "songs", :force => true do |t|
+  create_table "songs", force: true do |t|
     t.string   "title"
     t.string   "genre"
     t.float    "length"
@@ -399,19 +397,19 @@ ActiveRecord::Schema.define(:version => 20140419104846) do
     t.datetime "analyzed_at"
   end
 
-  create_table "songs_artists", :id => false, :force => true do |t|
+  create_table "songs_artists", id: false, force: true do |t|
     t.integer "song_id"
     t.integer "artist_id"
   end
 
-  create_table "songs_users", :id => false, :force => true do |t|
+  create_table "songs_users", id: false, force: true do |t|
     t.integer "user_id"
     t.integer "song_id"
   end
 
-  add_index "songs_users", ["user_id", "song_id"], :name => "by_song_and_user", :unique => true
+  add_index "songs_users", ["user_id", "song_id"], name: "by_song_and_user", unique: true
 
-  create_table "sources", :force => true do |t|
+  create_table "sources", force: true do |t|
     t.integer  "user_id"
     t.integer  "configuration_id"
     t.string   "type"
@@ -421,7 +419,7 @@ ActiveRecord::Schema.define(:version => 20140419104846) do
     t.datetime "updated_at"
   end
 
-  create_table "translations", :force => true do |t|
+  create_table "translations", force: true do |t|
     t.integer  "language_id"
     t.integer  "translatee_id"
     t.integer  "user_id"
@@ -430,37 +428,37 @@ ActiveRecord::Schema.define(:version => 20140419104846) do
     t.datetime "updated_at"
   end
 
-  create_table "users", :force => true do |t|
-    t.string   "email"
-    t.string   "encrypted_password",     :limit => 128, :default => "",    :null => false
+  create_table "users", force: true do |t|
+    t.string   "email",                  default: "",    null: false
+    t.string   "encrypted_password",     default: "",    null: false
     t.string   "reset_password_token"
     t.datetime "reset_password_sent_at"
     t.datetime "remember_created_at"
-    t.integer  "sign_in_count",                         :default => 0
+    t.integer  "sign_in_count",          default: 0
     t.datetime "current_sign_in_at"
     t.datetime "last_sign_in_at"
     t.string   "current_sign_in_ip"
     t.string   "last_sign_in_ip"
     t.string   "password_salt"
-    t.integer  "failed_attempts",                       :default => 0
+    t.integer  "failed_attempts",        default: 0
     t.string   "unlock_token"
     t.datetime "locked_at"
     t.datetime "created_at"
     t.datetime "updated_at"
-    t.boolean  "admin",                                 :default => false
+    t.boolean  "admin",                  default: false
     t.string   "image"
     t.string   "name_source"
-    t.boolean  "has_password",                          :default => true
+    t.boolean  "has_password",           default: true
     t.string   "custom_name"
-    t.string   "show_ads",                              :default => "all"
+    t.string   "show_ads",               default: "all"
     t.string   "unique_token"
   end
 
-  add_index "users", ["email"], :name => "index_users_on_email", :unique => true
-  add_index "users", ["reset_password_token"], :name => "index_users_on_reset_password_token", :unique => true
-  add_index "users", ["unlock_token"], :name => "index_users_on_unlock_token", :unique => true
+  add_index "users", ["email"], name: "index_users_on_email", unique: true
+  add_index "users", ["reset_password_token"], name: "index_users_on_reset_password_token", unique: true
+  add_index "users", ["unlock_token"], name: "index_users_on_unlock_token", unique: true
 
-  create_table "votes", :force => true do |t|
+  create_table "votes", force: true do |t|
     t.integer  "user_id"
     t.integer  "translation_id"
     t.integer  "value"
