@@ -5,8 +5,8 @@
 // the compiled file.
 //
 //= require jquery
+//= require jquery-ui
 //= require jquery_ujs
-//= require jquery.ui.all
 //= require jquery.qtip.min
 //= require jquery.color
 
