@@ -2,14 +2,25 @@
 class DeviseCreateUsers < ActiveRecord::Migration
   def change
     create_table(:users) do |t|
-      t.database_authenticatable :null => false
-      t.recoverable
-      t.rememberable
-      t.trackable
-      t.encryptable
-      #t.confirmable
-      t.lockable :lock_strategy => :failed_attempts, :unlock_strategy => :both
-      # t.token_authenticatable
+      t.string :email,              :null => false, :default => ""
+	  t.string :encrypted_password, :null => false, :default => ""
+	  
+	  t.string :reset_password_token
+	  t.datetime :reset_password_sent_at
+	  
+	  t.datetime :remember_created_at
+	  
+	  t.integer  :sign_in_count, :default => 0
+      t.datetime :current_sign_in_at
+      t.datetime :last_sign_in_at
+      t.string   :current_sign_in_ip
+      t.string   :last_sign_in_ip
+	  
+	  t.string :password_salt
+	  
+	  t.integer  :failed_attempts, :default => 0
+	  t.string   :unlock_token
+	  t.datetime :locked_at
       t.timestamps
     end
 
