@@ -1,5 +1,5 @@
 # -*- encoding : utf-8 -*-
-class AddConfigurationsToDevices < ActiveRecord::Migration
+class AddConfigurationToDevices < ActiveRecord::Migration
   def change
     add_column :devices, :configuration_id, :int
   end
