@@ -3,7 +3,7 @@ class CreateAdminConfigs < ActiveRecord::Migration
   def change
     create_table :admin_configs do |t|
       t.string :name
-      t.integer :pending_donations_limit
+      t.integer :pending_donations_limit, :default => 10
 
       t.timestamps
     end
