@@ -1,5 +1,5 @@
 class AddIndexUniqueEmailInUsers < ActiveRecord::Migration
   def change
-    add_index :users, :email, :unique => true
+    #add_index :users, :email, :unique => true
   end
 end
